refactor: updated project card structure
