+++ src/App.tsx
@@ -5,7 +5,6 @@ import { Footer } from "./components/Footer/Footer"
 import { Header } from "./components/Header/Header"
 import { Navbar } from "./components/Navbar/Navbar"
 import { Projects } from "./components/Projects/Projects"
-import { Skills } from "./components/Skills/Skills"
 import { ThemeProvider } from "./context/ThemeProvider"
 
 export const App = () => (
@@ -15,8 +14,7 @@ export const App = () => (
     <Header/>
     <AboutMe/>
     <Experience/>
-    {/* <Skills/> */}
-    {/* <Projects/> */}
+    <Projects/>
     {/* <Contact/> */}
     {/* <Footer/> */}
   </ThemeProvider>
