+++ src/assets/icons/LinkIcon.tsx
@@ -1,0 +1,9 @@
+export const LinkIcon = () => {
+  return (
+    <svg className="button-icon" width="48px" height="48px" xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 48 48">
+      <path className="cls-1" 
+        d="M44.3,21.7l-9,9c-5,5-13,5-18,0-.8-.8-1.4-1.7-1.9-2.6l4.2-4.2c.2-.2.4-.3.7-.5.3,1,.8,1.9,1.6,2.7,2.5,2.5,6.5,2.5,9,0l9-9c2.5-2.5,2.5-6.5,0-9-2.5-2.5-6.5-2.5-9,0l-3.2,3.2c-2.6-1-5.4-1.3-8.1-.9l6.8-6.8c5-5,13-5,18,0,5,5,5,13,0,18ZM20.4,36.6l-3.2,3.2c-2.5,2.5-6.5,2.5-9,0-2.5-2.5-2.5-6.5,0-9l9-9c2.5-2.5,6.5-2.5,9,0,.8.8,1.3,1.7,1.6,2.7.2-.1.5-.2.7-.4l4.2-4.2c-.5-.9-1.2-1.8-1.9-2.6-5-5-13-5-18,0L3.7,26.3c-5,5-5,13,0,18,5,5,13,5,18,0l6.8-6.8c-2.7.4-5.5.1-8.1-.9Z"
+      />
+    </svg>
+  )
+}
