--- src/components/DevProjectCard/DevProjectCardStyles.css
@@ -1,108 +1,0 @@
-.project-card {
-  border-radius: 1rem;
-  width: 20rem;
-  height: 30rem;
-  overflow: hidden;
-  margin: 1.5rem;
-  background-color: var(--background-elements);
-  box-shadow: var(--shadow);
-}
-
-.project-img-container {
-  width: 100%;
-  height: 12rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-  position: relative;
-
-  & .project-img {
-    width: 100%;
-    object-fit: fill;
-    object-position: center center;
-  }
-}
-
-.project-info-container {
-  display: flex;
-  width: 100%;
-  height: 18rem;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: flex-start;
-  padding: 1.25rem;
-
-  & .project-info {
-    display: flex;
-    flex-direction: column;
-
-    & .project-title {
-      margin-bottom: .8rem;
-      font-weight: var(--fira-extrabold);
-      color: var(--rock-red);
-    }
-
-    & .project-description {
-      width: 100%;
-      height: 5rem;
-      font-size: .85rem;
-    }
-  }
-}
-
-.tags-container {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-
-  & .tag-name {
-    font-size: .8rem;
-    font-weight: var(--fira-medium);
-    border-radius: .5rem;
-    color: #000;
-    background-color: #ebebeb;
-    margin: .25rem;
-    padding: .25rem .55rem;
-  }
-}
-
-.links-container {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-}
-
-@media screen and (max-width: 530px) {
-
-  .project-card {
-    width: 18rem;
-    height: 35rem;
-    margin: 1rem;
-  }
-
-  .project-img-container {
-    height: 10rem;
-  }
-
-  .project-info-container {
-    height: 25rem;
-
-    & .project-info .project-title {
-      font-size: var(--hero-text-size);
-    }
-
-    & .project-info .project-description {
-      height: 8rem;
-      font-size: var(--links-size);
-    }
-  }
-
-  .links-container {
-    flex-direction: column;
-
-    & .link-container {
-      margin: .5rem 0;
-    }
-  }
-}
