+++ src/components/ProjectCard/ProjectCardStyles.css
@@ -1,0 +1,41 @@
+.project-card {
+  width: 100%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  border: 1px solid red;
+}
+
+.img-container {
+  width: 27.125rem;
+  height: 15rem;
+  
+  & img {
+    width: 100%;
+    border-radius: 1rem;
+  }
+}
+
+.separator {
+  height: 65%;
+  border-left: 1px solid var(--border-color);
+}
+
+.project-content {
+  width: 32.75rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+
+  & .project-title{
+    font-size: var(--subtitle-size);
+    font-weight: var(--outfit-semi-bold);
+  }
+}
+
+
+@media screen and (max-width: 530px) {
+
+}
