+++ src/components/Projects/ProjectsStyles.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding: 2.5rem 5rem;
+  padding: 3rem 0;
 }
 
 .project-cards-container {
