--- src/components/SkillIcon/SkillIconStyles.css
@@ -1,24 +1,0 @@
-.icon-container {
-  width: 3.5rem;
-  height: 3.5rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 1rem;
-
-  & .img-container {
-    width: 3rem;
-
-    & .icon-img {
-      width: 100%;
-      object-fit: fill;
-      object-position: center center;
-    }
-  }
-
-  & .icon-name {
-    font-size: .8rem;
-    font-weight: var(--fira-semi-bold);
-  }
-}
