--- src/components/Skills/SkillsStyles.css
@@ -1,58 +1,0 @@
-.skills-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 2.5rem 5rem;
-}
-
-.skills-icons-container {
-  width: 37.5rem;
-  padding: 2rem;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-
-
-  & .development-skills-container {
-    width: 17rem;
-  }
-
-  & .design-skills-container {
-    width: 15rem;
-  }
-
-  & .skills-title {
-    font-weight: var(--fira-bold);
-  }
-
-  & .icons-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    margin: 1rem 0;
-  }
-}
-
-@media screen and (max-width: 830px) {
-
-  .skills-icons-container {
-    width: 100%;
-  }
-  
-}
-
-@media screen and (max-width: 600px) {
-
-  .skills-container {
-    padding: 2.5rem;
-  }
-
-  .skills-icons-container {
-    flex-direction: column;
-    align-items: center;
-
-    & .design-skills-container {
-      margin-bottom: 3rem;
-    }
-  }
-}
